pyproject.toml
[project]
name = "agent-social-media"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "crewai-tools",
    "crewai>=0.100.0",
    "flask>=3.1.0",
    "requests>=2.32.3",
    "pillow>=11.1.0",
    "evolutionapi>=0.0.9",
    "pilgram>=1.2.1",
    "imgurpython>=1.1.7",
    "google-generativeai>=0.8.4",
    "streamlit>=1.32.0",
    "onnxruntime>=1.17.0",
    "python-dotenv>=0.19.0",
]


